fix: move navigation effect to profile page
