+++ packages/apps/human-app/frontend/src/modules/worker/profile/types/profile-types.ts
@@ -1,5 +1,4 @@
 export interface WorkerProfileStatus {
-  emailVerified: boolean;
   kycApproved: boolean;
   kycDeclined: boolean;
   kycToComplete: boolean;
